No doctest
When I run `cargo test`, I see lines for [doc]tests I do not have which annoyed me for a while.
I also add a cargo config for my target directory to be on a SSD for faster compile times, but no need for it to be public so I gitignore it.