Do not share inputs publicly
See https://www.reddit.com/r/adventofcode/wiki/faqs/copyright/inputs/

That was obvious to me that copy html pages and share them was not okay but I did not thought that sharing inputs was not allowed, sorry the gitignore file is fixed.

I'm gonna restore the inputs from the trash after this. It seems that changing the gitignore file is not enough to untrack files. If you know better, please tell me.