aoc-macro: auto-detect dependencies
Instead of providing numbers representing a range of "aoc**-**" dependencies, we provide the path to the project.
Then we can parse the manifest "Cargo.toml" thanks to `cargo_toml` and automatically detect the wanted dependencies.

So now, there is no need to update the macro. Add a dependency to the manifest is enough. And compile times remain very good.

I liked the previous version, I learnt a lot. But this is just better.