2022 day 7
Today, I first used LsEntry and Command with strings because I failed again with lifetimes.
But I learnt that is because the trait `FromStr` just does not accept lifetimes at all (it's clear from the documentation: "no lifetime parameter").
Instead, I now use `TryFrom<&'a str>` and it works fine.