New utils: char10, char16
To convert a char that represents a decimal/hexadecimal digit to u32 (or any type that implements `TryFrom<u32>`.
It shorten some previous solvers even if we sometimes need to explicit the type we want (because we can't give a default type defining these utils).
I thought about creating `digit_char<const N: u32, T>` but it would require to write the turbofish notation each time to only give 10 (or rarely 16).