New trait: OkIterator
I do not unwrap results to not panic, so I have a bunch of iterators of results.
The previous trait "FromIterStr" solved most of the struggle, but not all of it. It was "map & collect to Result" from "iterators of str".
For some time, I felt that "map" should not be mixed up with "collect" and I wanted more than just collect from iterators of results (also sum, product...) so a trait for those make more sense. It adapts most Iterator's methods when it makes sense to me.
At first, I did not set the error type to be `anyhow::Error` but it would require too many times to indicate that it is even when it is obvious to me. So I set it to be `anyhow::Error` and therefore I must handle iterators of other result types differently with `str_parse`: `str::parse` but returning the error type I need.
And I changed `parse_to_grid[_with_loc]` to be functions in order to delete the old trait `FromIterStr`.