Times for a whole year & Speed up 2022 day 24
"cli-runner" can now display times for all puzzles of a given year.

"aoc22-24" took a lot of times to parse. I now use a vector of booleans instead of a hashset because it's faster, and I independently consider rows and columns which significantly reduces the length of the vectors and speed up the whole process.

Except for "aoc22-19" that relies on an external program (20.9 seconds) that I can't optimize without replacing it entirely, all puzzle of 2022 are solved in 769ms.
And 8.5 seconds for 2021, which is not good enough for my taste yet.