aoc21-19: twice faster
aoc22-23: A little bit faster with a slice instead of an array. Not sure why.

aoc21-19: Thanks to a "flamegraph", I know that 95% of the time was spent in the "counts" method, and 51% internally in HashMap's method "reserve". So here by setting a capacity, we do not reallocate and only spend half the time.
I also early stop counting when a count 12 is reached but it does not really save time.