aoc21-23: 16 times faster
This was a mess so I started from nearly scratch. It was the longest program for 2021.
Thanks to a flamegraph, I know the bottleneck was computing the energy to the goal and more specifically getting the distance from a hash map.
Now, all paths and distances are precomputed and stored in a 2D grid.
The struct "State" is also minimalist, but it remains nicely index-able.