aoc21-08: 22 times faster
Use owned strings and hashs was slow. The approach is identical (still 5040 permutations to consider) but use `u8` to represent a seven segment display is definitely lighter.
And use `u128` to represent an unordered collection of `u8` is a bit tricky (and probably not as essential to speed, sort vectors of 10 integers can not be slow) but I like it.
I also added the simple example to tests.