2020 day 19
I first did part 1 in python with regexps (second day falling back to python).
I ended up using "pest" again in rust to solve part 2 which needed to parse the grammar at compile time, so the input is manually manipulated. And it was not particularly nice this time, because it was not the right tool for the job, just like regexps.
So I changed my strategy again and tried what I should have done from the beginning: backtrack.
What a day! I now understand the stats of this day I saw on the website.
Next: day 20, I'm in danger!