aoc20-15: 60% faster
This solution is by far the slowest of AoC 2020: 1.16 second now decreased to 0.7 second.
Use u32 instead of usize. Use [u32, u32] instead of Option<(u32, u32)> with a default value of [0, 0]. This default value defeats the whole point of using options but I guess that with 30 millions of them, it's worth it.