Remove "build.target-dir" from the cargo config
At first, I only used one computer for this project but now I sometimes change (for a flamegraph on Linux or on a laptop for the week end) and then need to remove this line first which I usually forgot.
Instead, I made a symlink from "/target" to my SSD. It's just better.