aoc22-19: Also solve without linear programming
I enjoyed solving it with `good_lp` as I learnt a lot with it but "Integer LP" is NP-hard so quite slow.
So now, I basically solve it with "branch and bound".
However, to not remove my first solution, I hide it behind a feature flag: "lp".
Test is down to 0.08 second from 20-30 seconds. And now all 2022 is solved under 0.7 second. 🤩