Add a "final" profile
I witnessed it be between 1% and 14% faster than release (depends on the year), roughly 9% faster total.
So this is will build my "production" binary. Therefore, flamegraph will profile this instead and the "cli" alias will use it and be finally useful.
I keep "release" as is to have fast enough compile times.