aoc21-19: 24 permutations, not 48
It's the currently slowest solution of 2021 and I tried many times to fasten this without addressing the elephant in the room: the 48 rotations when 24 should be enough. I finally do!
After some struggle, I came up with what I think is an elegant descriptive solution. Which axis is `Z` ? Is it up or down? And how rotated is it around it.
Now all 2021 runs around 775ms, finally under 1s. 🤩