Use `HashMap/HashSet` from `rustc-hash` instead of `std::collections`
I first thought of putting this in the `utils` library but I don't want to create a prelude in it, and I consider that a prelude is necessary to be able to use `FxHasherHack` without naming it every time as it would defeat the all point of a (nearly) "drop-in replacement". That's why it's in `common`.

I did not thought I used `HashMap::from` but I did. Use `insert` for a single element or `.into_iter().collect()` otherwise.
Another struggle I anticipated is that I can not make "itertools" use "rustc-hash", (don't want to fork it for this) and types might be different but it only happened once in "aoc20-20".

Finally with clippy, I disallow `HashMap/HashSet` types from `std::collections`.
I must mention/thank this great ressource: https://nnethercote.github.io/perf-book/hashing.html

Nextest runs 10% faster (including solvers that do not use hashes). And AoC 2020 finally runs just under one second: 990ms. 🤩

Update README.md