aoc15-09 & aoc15-13: 5 times faster
`permutohedron` is a nice alternative to process all permutations but its default iterator allocate each item on the heap but that's easily fixable since we just need to map by a function which can use the slice without heap allocation.
aoc21-08 is not really faster but aoc15-09 and aoc15-13 which were already fast are even more.