ureq without cc: native-tls instead of rustls
"ureq" uses "rustls" by default which requires "cc" to be build. Therefore, it requires an installed C compiler.
It was not the case for me last recently on another computer and because I did not want to install "gcc" I ended up making a branch without my own "web" utility for the time being.
This uses "native-tls" instead.