Bench my solvers on various inputs
Before I changed the folder structure, I had this in a separate branch of this git repository, except it was based on a `paste` macro and `criterion` macros. But after expanding the `criterion_group` and `criterion_main` macros, I realized this would be so much simpler without those macro.

And I can benchmark against other inputs as well.